Rename io to output
